--- src/main/java/com/space/server/chat/domain/value/Type.java
@@ -1,6 +1,0 @@
-package com.space.server.chat.domain.value;
-
-public enum Type {
-    CODE,
-    HINT
-}
