--- src/main/java/com/space/server/chat/presentation/dto/request/CreateChatRequest.java
@@ -1,8 +1,0 @@
-package com.space.server.chat.presentation.dto.request;
-
-import com.space.server.chat.domain.value.Type;
-
-public record CreateChatRequest(
-        String userChat,
-        Type type
-) {}
