+++ src/main/java/com/space/server/domin/chat/domain/value/Type.java
@@ -1,0 +1,6 @@
+package com.space.server.domin.chat.domain.value;
+
+public enum Type {
+    CODE,
+    HINT
+}
