+++ src/main/java/com/space/server/domin/chat/presentation/dto/request/CreateChatRequest.java
@@ -1,0 +1,8 @@
+package com.space.server.domin.chat.presentation.dto.request;
+
+import com.space.server.domin.chat.domain.value.Type;
+
+public record CreateChatRequest(
+        String userChat,
+        Type type
+) {}
