+++ src/main/java/com/space/server/domin/chat/service/QueryChatService.java
@@ -1,8 +1,8 @@
-package com.space.server.chat.service;
+package com.space.server.domin.chat.service;
 
-import com.space.server.chat.domain.Chat;
-import com.space.server.chat.presentation.dto.response.ChatResponse;
-import com.space.server.chat.service.implementation.ChatReader;
+import com.space.server.domin.chat.domain.Chat;
+import com.space.server.domin.chat.presentation.dto.response.ChatResponse;
+import com.space.server.domin.chat.service.implementation.ChatReader;
 import com.space.server.domin.quiz.service.implementation.QuizReader;
 import com.space.server.state.service.implementation.StateReader;
 import com.space.server.user.service.implementation.UserReader;
